+++ NEWS.md
@@ -1,3 +1,20 @@
+# shapr 1.0.5
+
+* Added `get_results()` function for extracting key components from a `shapr` object, including Shapley values, 
+model parameters, and iterative computation details ([#460](https://github.com/NorskRegnesentral/shapr/pull/460)).
+* Added `summary.shapr()` method, which uses `get_results()` and internal `format_xyz()` functions to provide a concise
+overview of the explanation object ([#460](https://github.com/NorskRegnesentral/shapr/pull/460)).
+* Updated `print.shapr()` to support printing of specific components with customizable formatting for digits and 
+decimal precision ([#460](https://github.com/NorskRegnesentral/shapr/pull/460)).
+* Updated vignettes, examples and tests to reflect the new functionality 
+([#460](https://github.com/NorskRegnesentral/shapr/pull/460)).
+* Refactored `cli`-based output functions for better reuse across multiple methods
+([#460](https://github.com/NorskRegnesentral/shapr/pull/460)).
+* A small range of other minor visual fixes ([#460](https://github.com/NorskRegnesentral/shapr/pull/460)).
+* Fixed error with many features by removing `as.integer()` coercion when computing number of coalitions per 
+coalition size ([#462](https://github.com/NorskRegnesentral/shapr/pull/462)).
+
+
 # shapr 1.0.4
 
 * Implement semi-deterministic sampling of coalitions similar to the default in the *shap* python library and described